fix(swiftgit/windows): cast status bitmask + remaining merge bitor to UInt32
The Windows compiler imports `git_status_t` with an `Int32` raw value;
our `(raw.status.rawValue & GIT_STATUS_*.rawValue)` checks only worked
on Apple/Linux. Funnel both sides through `UInt32(...)`. Same fix for
the `GIT_CHECKOUT_ALLOW_CONFLICTS` bit-or in Merge that I missed in
the previous sweep.

+++ Sources/SwiftGit/Lib/GitClient+Merge.swift
@@ -187,7 +187,7 @@ extension GitClient {
         var coOpts = git_checkout_options()
         try check(git_checkout_options_init(&coOpts, UInt32(GIT_CHECKOUT_OPTIONS_VERSION)))
         coOpts.checkout_strategy = UInt32(GIT_CHECKOUT_SAFE.rawValue)
-            | GIT_CHECKOUT_ALLOW_CONFLICTS.rawValue
+            | UInt32(GIT_CHECKOUT_ALLOW_CONFLICTS.rawValue)
 
         var heads: [OpaquePointer?] = [theirAC]
         _ = try heads.withUnsafeMutableBufferPointer { buf in
